requirements.txt
Django>=1.8
dj-database-url==0.3.0
django-extensions==1.4.6
gunicorn==19.2.1
coverage==3.7.1
django-coverage==1.2.4
boto==2.36.0
psycopg2>=2.5.1
unholster.milieu==1.0.0
werkzeug==0.9.6
django-password-reset==0.7
django-bootstrap-form==3.1
-e git+ssh://git@github.com/unholster/django-frontflow.git#egg=django-frontflow
python-social-auth==0.2.12
githubpy==1.1.0
